Part 3 - Model Serving microservice/requirements.txt
# production
pyyaml
numpy
pandas
scikit-learn
boto3
metaflow
flask
gunicorn
tensorflow-hub
tensorflow-datasets

# dev
coloredlogs
black
flake8

# etc
